job:#10054 Disconnected and reconnected interfaces
The C_SF instances in the GPS_Watch system diagram had a value for
C_SF.Label (after a load and persist) that looked like this:
[undetermined path] -o)- [undetermined path]

Of course this is after I had moved the components into separate
projects. I disconnected and reconnected the interfaces, and it resolved
this problem.