job:#10054 Saved changed from Load and Persist in informal diagrams
These are changes I have been avoiding making because it is the spot is
graphics where the tool is not persisting the full path to the
GD_GE.represents_path in some cases. However, the GD_MD does have the
full path, and thus the GD_GE should not need one, and, as stated in
prior comments, there is no indication that this is a problem though
issue #10126 has been raised to track it.