job:#10054 removed interface Location::shared::LoctionUtil
This leaves the Location domain with a single interface. I moved the 1
and only operation that was in that interface into interface
Location::shared::LocationProvider. I also moved the action body that
was in Location:UTIL port into the Location:LOC port (port Location:UTIL
was removed when the interface was deleted). I fixed up references to
this operation accordingly too.