job:#10054 I am checking-in the GPS_Watch Java class binaries
Comment from a community post about this I made on Mar 16, 2018:  I am
having an Eclipse/Java project build dependency issue... I reworked the
GPS model and separated it into multiple projects (separate projects for
each domain). One of the domains is UI. In GPS, this domain, along with
the UI model, holds a java gui. This project therefore has a java
nature. The problem is with dependencies. The issue is that the build is
not happening unles the java source if modified, so when the bin folder
(where java classes are output on build) is cleaned, the build does not
happen unless the java source is modified. Looking at the project before
I changed it, I see that the "workaround" that was put in place for this
was to add the c "src" folder to the java classpath. This caused the
java build to happen when the model was built. That was a hack that
should not have been present. I don't do enough java peoject build to
see why this project is not dependent on the bin folder. Can someone
help me?
I am tempted to just check in the .class files. I don't like putting
binaries in revision control, but I don't want to spend time on this...
I think that is what I will do.