job:#10054 Create a GPS_Watch::GPS_Watch component
I modified the GPS_Watch system component diagram and put it in the
shape that facilitates all communication going through the component. I
also introduced an new interface, Tracking, in the Tracking component,
and I created a provided interface on the Tracking component and
formalized it to this new interface. I note that the graphics of the
component references to Tracking were not updated. I will deal with this
bug in the next step (after this current step is tested).

I still did not do anything that should affect the current running
model's wiring. This is simply another incremental step.