job:10054 Wired the new Domain interfaces added in the prior step
I wired the domain's required interface to the project's provided
interface. I also renamed the port in the GPS_Watch project to 
<domain>__<terminator> (2 underscores) as required by the convention
described in the "MASL Modeling and Conversion Guide". I will note that
I did NOT perform this action on :
HeartRateMonitor::Shared::HeartRateMonitor_Unspecified

The reason is that the required interface for this did NOT get created
on the GPS_Watch component diagram. I will resolve this problem in a
stand-alone incremental step next (after this current step is first
tested)