job:#10054 GPS_Watch component diagram has been rewired (not executing)
I disconnected the old GPS Watch wiring that was between domains and
connected the new wiring through the component. I had to move some of
the OAL from required interface ports to provided interface ports, and
in doing I changed the message direction to "to provider".

The state with this change is that the start/stop pressed button message
is being delivered to the UI, but the other domains are not running
after that and the UI is not updated.