job:#10054 Fixed problems found in CSM to ISM conversion
This was mostly in the Location component, but others too. The
initialization was not correct for ISM operation and the component was
being deactivated before it ever executed. I also removed some event to
self that were in the "tick" states in favor of having the tick event
just be generated to self as it was before and the running/idle state
handles the tick.