job:#10054 There is a bug in Verifier's handling of realized EEs
Verifier does not seem to be handling the situation where there is a
multi-domain project where there are realized EEs in multiple domains.
The situation in this case is that execution hangs in the Location
component when calling out to MATH::sqrt.  I have tried multiple
senarios to deal with this including moving the MATH.java class into
different domains, I have not been able to get any of the things I tried
to work. For now what I did is I left MATH.lib where it originally was
(in the UI project). I THINK it should really be in the Location project
where it is actually used, but that does not work either (same hang in
verifier). Note that the EE that defines the sqrt operation is in
GPW_Watch. I tried to put it there too (same result).

To work around this I introduced a domain function Location::sqrt that
calculates a sqrt (approximation). I also reopened issue
https://support.onefact.net/issues/7658 that, I believe is where we
should be testing this.